Remove Runner contract Lab dependency
