+++ crates/contracts/homeboy-api-jobs-contract/src/types.rs
@@ -5,8 +5,9 @@ use serde_json::Value;
 use uuid::Uuid;
 
 use crate::metadata::JobArtifactMetadata;
-use homeboy_lab_contract::path_materialization::PathMaterializationPlan;
-use homeboy_runner_contract::{RunnerJobLifecycleMetadata, RunnerLifecycleOwner};
+use homeboy_runner_contract::{
+    PathMaterializationPlan, RunnerJobLifecycleMetadata, RunnerLifecycleOwner,
+};
 use homeboy_source_snapshot_contract::SourceSnapshot;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
