+++ crates/contracts/homeboy-lab-contract/Cargo.toml
@@ -17,3 +17,4 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 homeboy-error = { version = "0.1.0", path = "../../homeboy-error" }
 homeboy-redaction = { version = "0.1.0", path = "../../homeboy-redaction" }
+homeboy-runner-contract = { version = "0.1.0", path = "../homeboy-runner-contract" }
