+++ crates/contracts/homeboy-lab-contract/src/lib.rs
@@ -1,10 +1,9 @@
 //! Shared Lab/runner contract types for homeboy.
 //!
 //! This crate holds the contract surface that both the core engine and the
-//! CLI/command-contract layer depend on: the Lab workload/handoff types plus the
-//! env/secret/path materialization plans and agent-task config/policy data they
-//! reference. Extracting it below both `core` and `command_contract` breaks the
-//! former `core <-> command_contract` dependency cycle.
+//! CLI/command-contract layer depend on: Lab workload/handoff and agent-task
+//! policy types. Generic runner materialization contracts are compatibility
+//! re-exports from `homeboy-runner-contract`.
 
 pub mod agent_task_config;
 pub mod agent_task_outcome;
@@ -16,6 +15,7 @@ pub mod path_materialization;
 pub mod secret_env_plan;
 
 pub mod lab {
+    pub mod execution_envelope;
     pub mod handoff;
     pub mod labels;
     pub mod transport_failure;
