+++ crates/contracts/homeboy-runner-contract/Cargo.toml
@@ -8,7 +8,7 @@ repository = "https://github.com/Extra-Chill/homeboy"
 publish = false
 
 [dependencies]
-homeboy-lab-contract = { version = "0.1.0", path = "../homeboy-lab-contract" }
 homeboy-source-snapshot-contract = { version = "0.1.0", path = "../homeboy-source-snapshot-contract" }
+homeboy-redaction = { version = "0.1.0", path = "../../homeboy-redaction" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
