+++ crates/homeboy-lab-runner/Cargo.toml
@@ -21,6 +21,7 @@ homeboy-source-snapshot-contract = { version = "0.1.0", path = "../contracts/hom
 homeboy-upgrade = { version = "0.1.0", path = "../homeboy-upgrade" }
 homeboy-rig = { version = "0.1.0", path = "../homeboy-rig" }
 homeboy-runner-contract = { version = "0.1.0", path = "../contracts/homeboy-runner-contract" }
+homeboy-lab-contract = { version = "0.1.0", path = "../contracts/homeboy-lab-contract" }
 homeboy-agents = { version = "0.1.0", path = "../homeboy-agents" }
 homeboy-fuzz = { version = "0.1.0", path = "../homeboy-fuzz" }
 homeboy-product-identity = { version = "0.1.0", path = "../homeboy-product-identity" }
